First cut at reintroducing object serialization.
This is very much based on the work t0m did in his json branch (which
was merged into swann to become swann-with-json). The nicely broken up
Git objects created more replicated code than I like but it seems to work
which is enough for now. Next up is the rest of the objects and then onto
a shiny new REST interface.